third_party/blink/web_tests/virtual/backface-visibility-interop/external/wpt/css/css-transforms/3d-rendering-context-behavior-expected.txt

This is a testharness.js-based test.
PASS Direct DOM parent is root of rendering context
PASS Intermediate DOM nodes cause rendering context to end (normal flow)
FAIL Intermediate DOM nodes cause rendering context to end (absolute) assert_equals: expected 100 but got 200
FAIL Intermediate DOM nodes cause rendering context to end (fixed) assert_equals: expected 100 but got 200
PASS Perspective applies to direct DOM normal-flow children
PASS Perspective applies to direct DOM abs-pos children
PASS Perspective applies to direct DOM fixed-pos children
PASS Perspective does not apply to DOM normal-flow grandchildren
PASS Perspective does not apply to DOM abs-pos grandchildren
PASS Perspective does not apply to DOM fixed-pos grandchildren
Harness: the test ran to completion.
